0.2.7-buildfix2
- added icon for the actual exe file (via .o file)
- python script to make the icon, and doing rc & windres stuff
- no need to redownload ninja if there's available

+++ tools/cmakeautomate_mingw32.py
@@ -41,34 +41,44 @@ folder_name = "build"
 build_folder_path = os.path.join(script_dir, folder_name)
 
 if os.path.exists(build_folder_path):
-    shutil.rmtree(build_folder_path)
+    # shutil.rmtree(build_folder_path)
+    for root, dirs, files in os.walk(build_folder_path, topdown=False):
+        for name in dirs:
+            shutil.rmtree(os.path.join(root, name))
+
+        for name in files:
+            if name != 'ninja.exe':
+                os.remove(os.path.join(root, name))
+    print("build directory cleaned.")
 
 os.makedirs(build_folder_path, exist_ok=True)
 
 # Define the URL to download and the filename
 url = "https://github.com/ninja-build/ninja/releases/download/v1.11.1/ninja-win.zip"
 filename = "ninja-win.zip"
+
 ninja_zip_path = os.path.join(build_folder_path, filename)
+ninja_exe_path = os.path.join(build_folder_path, "ninja.exe")
 
 unzip_command = f"powershell Expand-Archive -Path {ninja_zip_path} -Destination {build_folder_path}"
-
-# Download the file using curl
 download_command = f"curl -L -o {ninja_zip_path} {url}"
-subprocess.call(download_command, shell=True)
 
-# Check if the download was successful
-if os.path.exists(ninja_zip_path):
-    # Extract the downloaded zip file
-    subprocess.call(unzip_command, shell=True)
+if not os.path.exists(ninja_exe_path):
+    # Download the file using curl
+    subprocess.call(download_command, shell=True)
 
-    # Clean up the zip file
-    os.remove(ninja_zip_path)
-    print("Done downloading ninja.")
-else:
-    print("Download failed.")
-    exit()
+    # Check if the download was successful
+    if os.path.exists(ninja_zip_path):
+        # Extract the downloaded zip file
+        subprocess.call(unzip_command, shell=True)
 
-ninja_exe_path = os.path.join(build_folder_path, "ninja.exe")
+        # Clean up the zip file
+        os.remove(ninja_zip_path)
+        print("Done downloading ninja.")
+    else:
+        print("Download failed.")
+        exit()
+
 mingw_path = os.path.join(script_dir, "mingw32")
 if os.path.exists(mingw_path):
     c_compiler_path = os.path.join(script_dir, r"mingw32\bin\gcc.exe").replace("\\","/")
